+++ CHANGELOG.rst
@@ -7,7 +7,11 @@ Release 2.0.6 (in development)
   camelCase attributes are generated too (however this will be dropped for
   py3k).
 
-* Renamed niftoaster spell fix_strip to modify_delblocks.
+* Renamed a few niftoaster spells.
+
+  - fix_strip -> modify_delblocks
+
+  - fix_disableparallax -> modify_disableparallax
 
 Release 2.0.5 (Nov 23, 2009)
 ============================